some norme fixed

+++ vis_init.c
@@ -6,7 +6,7 @@
 /*   By: uhand <uhand@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/18 15:27:28 by uhand             #+#    #+#             */
-/*   Updated: 2019/09/18 18:26:42 by uhand            ###   ########.fr       */
+/*   Updated: 2019/09/18 21:43:26 by uhand            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -71,8 +71,8 @@ void		coord_init(t_intldta *indta, t_graph *g)
 		g->scale--;
 		g->delta_x = (g->max_x - g->min_x) * g->scale;
 		g->delta_y = (g->max_y - g->min_y) * g->scale;
-		c.dx = ((long long)g->max_x - (long long)g->min_x) * (long long)g->scale;
-		c.dy = ((long long)g->max_y - (long long)g->min_y) * (long long)g->scale;
+		c.dx = ((LL)g->max_x - (LL)g->min_x) * (LL)g->scale;
+		c.dy = ((LL)g->max_y - (LL)g->min_y) * (LL)g->scale;
 	}
 	if ((!g->scale && g->scale_rec < 2) || (indta->r_flag && !redef))
 	{
